feat: building list of clients

+++ src/pages/Login/index.tsx
@@ -2,25 +2,45 @@ import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import api from '../../services/api';
 import * as S from './styles';
+import Swal from 'sweetalert2';
 
 function Login() {
   const history = useHistory();
   const [userName, setUserName] = useState('');
+  const [loginUser, setLoginUser] = useState(false);
 
   // useEffect(() => {
-  // }, []);
+  //   // login = 'true';
+  // }, [loginUser]);
+
+  function toast(icon: any, msg: any) {
+    Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      showConfirmButton: false,
+      timer: 5000,
+      timerProgressBar: true,
+    }).fire({
+      icon: icon,
+      title: msg
+    });
+  }
 
   async function handleSubmit() {
     await api.get('/clients')
       .then(async (response: any) => {
         if (response && response.data) {
           response.data.map((user: any) => {
-            console.log(user)
             if (user.name == userName) {
+              setLoginUser(true)
               localStorage.setItem("token", '1')
-              localStorage.setItem("user", '1')
+              localStorage.setItem("user", userName)
+              localStorage.setItem("fullUser", JSON.stringify(user))
+              localStorage.setItem("clientList", JSON.stringify(response.data))
               history.push('/home')
-              console.log(history)
+              document.location.reload()
+            } else{
+              toast('error', 'Erro ao realizar login!');
             }
           })
         }
